commandline/pwd - improve English language
Improve the English language in
en/commandline/printworkingdirectorypwdcommand.json.

In this commit, I have preserved as much of the original voice of the
author as possible while breaking run-on sentences into shorter,
more-easily-parseable sentences. Throughout this project, it will be my
intention to preserve as much of the original authorial voice as the
limits of English grammar will permit.

Signed-off-by: Zac Dover <zac.dover@proton.me>